Add a WebAssembly module wrapped as Web Component
The C++ code added by this commit is designed to be compiled into a
WebAssembly using Emscripten 2.0.16. Use the "build.sh" bash script to
do so.

The file "color-triangle-post.js" wraps the resulting WebAssembly module
as a Web Component by using the Custom Elements API (see [1]). That Web
Component is used by the file "color-triangle.html" and should display
a triangle filled with red, green and blue when opened in a web browser.

[1] https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_custom_elements