Make the color of the triangle changeable
This commit modifies the existing code so that the fill color of the
rendered triangle can be changed. To do so, it expands the interface of
the Web Component with a "red", "green" and "blue" HTML attribute and
corresponding JavaScript properties.

In addition, the HTML file is expanded with three sliders that modify
the values of those attributes / properties.