Send an event if the triangle color changes
Whenever the fill color of the triangle gets changed, either by
modifying the HTML attributes or the JavaScript properties, an event
gets fired. The event contains the new color of the triangle which gets
printed to the web browser's console.

Note that the closure compiler that is running as part of the build is
unaware of the JavaScript callback function that is executed by the C++
code. That is because that callback function is provided using emcc's
--extern-post-js switch.

In order to prevent the closure compiler from pointing out that this
callback function is missing it is necessary to provide this function as
an external symbol. This is done by adding a so called "externs" file.
That necessary file is added by this commit and submitted to emcc using
the EMCC_CLOSURE_ARGS environment variable.