Identify the instance that triggered an event
The JavaScript callback function that gets invoked by the C++ code is
just a hard-coded function that is not related to the Web Component
class. In order to identify within that callback function which Web
Component instance triggered an event, this commit passes the C++ "this"
pointer as argument to that function.

The callback function sends the "this" pointer forward as the "sender"
argument of the event. Finally, the Web Component code that reacts on
that event is able to determine whether the event was caused by its own
color triangle WebAssembly instance or not.

In order to make this work it is necessary to prevent the closure
compiler from renaming "$$.ptr". This is done by adding the
corresponding symbols to the externs file.