Attach a shadow DOM to hide the canvas
The intention behind adding a shadow DOM to the Web Component is to hide
its "internals". Specifically in this case the canvas. This however has
the side effect that the JavaScript code generated by Emscripten can no
longer access the canvas.

Therefore, the WebGL context cannot be created which results in a
runtime error. To be able to diagnose this better, this commit disables
the closure compiler which would otherwise obfuscate the name of the
WebGL context.

There isn't a solution available for this yet. Turning on the shadow DOM
currently just breaks things.