Create a new module for each Web Component instance
In order to support multiple instances of the Web Component it is a
necessity that each Web Component instance contains its own canvas
element. To be able to communicate to the WebAssembly module which
canvas instance to work with this commit creates a new module for each
instance of the Web Component.

Registering the canvas element with the Module object that is passed to
the factory function that creates the WebAssembly module instance
enables this. However, the downside here is that the lifecycle of the
module is now connected to the lifecycle of the Web Component which
means that creating a new Web Component will again fetch the .wasm file
from the web server.

This is unnecessary because a local copy of the .wasm file already
exists if another Web Component instance already exists. Hence, that
approach is inefficient but at the time of this commit no other approach
was found to support multiple Web Component instances.

Another downside is that the approach to identify the Web Component
instance that triggered an event which was introduced in commit [1] no
longer works. Creating the instance of the ColorTriangle class results
in the same address even though they reside in two differen WebAssembly
modules. Hence, distinguishing ColorTriangle instances by addresses no
longer works.

[1] 5898ed4d7c77f9c4fcc41476c366e6cfc6cacfce